[7.0] Fetch *first* docs (instead of latest) from .monitoring-kibana-* (#207) (#210)
This is because the kibana_stats.usage is not collected every collection any more. It is collected in the very first collection and then once every day after that. So for the most thorough comparison we now fetch the first docs from `.monitoring-kibana-*` (natively collected docs)  and `.monitoring-kibana-*mb-* (metricbeat collected docs) since these must contain the `kibana_stats.usage` field in them.